update Sentry setup / add do nothing to create translations
